create separate Auth class #18
